api/requirements.txt
azure-functions
azure-identity
azure-ai-projects
azure-ai-inference
